+++ GRDB/RowModel.swift
@@ -37,6 +37,15 @@ methods that define their relationship with the database:
 */
 public class RowModel {
     
+    /// The result of the RowModel.delete() method
+    public enum DeletionResult {
+        /// A row was deleted.
+        case RowDeleted
+        
+        /// No row was deleted.
+        case NoRowDeleted
+    }
+    
     /// A primary key. See RowModel.databaseTable and Table type.
     public enum PrimaryKey {
         
@@ -234,16 +243,24 @@ public class RowModel {
     
     On successful insert, this method sets the *edited* flag to false.
     
+    This method is guaranteed to have inserted a row in the database if it
+    returns without error.
+    
     - parameter db: A Database.
     - throws: A DatabaseError whenever a SQLite error occurs.
     */
     public func insert(db: Database) throws {
-        try withDataMapper { dataMapper in
-            let insertionResult = try dataMapper.insert(db)
+        try withDataMapper { dataMapper -> Void in
+            let changes = try dataMapper.insert(db)
             
             // Update RowID column if needed
-            if let (rowIDColumn, insertedRowID) = insertionResult {
-                setDatabaseValue(DatabaseValue.Integer(insertedRowID), forColumn: rowIDColumn)
+            if let primaryKey = self.dynamicType.databaseTable?.primaryKey, case .RowID(let rowIDColumn) = primaryKey {
+                guard let rowID = dataMapper.storedDatabaseDictionary[rowIDColumn] else {
+                    fatalError("\(self.dynamicType).storedDatabaseDictionary must return the value for the primary key `(rowIDColumn)`")
+                }
+                if rowID == nil {
+                    setDatabaseValue(DatabaseValue.Integer(changes.insertedRowID!), forColumn: rowIDColumn)
+                }
             }
             
             edited = false
@@ -255,6 +272,9 @@ public class RowModel {
     
     On successful update, this method sets the *edited* flag to false.
     
+    This method is guaranteed to have updated a row in the database if it
+    returns without error.
+    
     - parameter db: A Database.
     - throws: A DatabaseError is thrown whenever a SQLite error occurs.
               RowModelError.RowModelNotFound is thrown if the primary key does
@@ -262,9 +282,8 @@ public class RowModel {
               updated.
     */
     public func update(db: Database) throws {
-        try withDataMapper { dataMapper in
+        try withDataMapper { dataMapper -> Void in
             try dataMapper.update(db)
-            
             edited = false
         }
     }
@@ -278,6 +297,9 @@ public class RowModel {
     Otherwise, performs an insert.
     
     On successful saving, this method sets the *edited* flag to false.
+    
+    This method is guaranteed to have inserted or updated a row in the database
+    if it returns without error.
     
     - parameter db: A Database.
     - throws: A DatabaseError whenever a SQLite error occurs, or errors thrown
@@ -304,16 +326,23 @@ public class RowModel {
     On successful deletion, this method sets the *edited* flag to true.
     
     - parameter db: A Database.
+    - returns: Whether a row was deleted or not.
     - throws: A DatabaseError is thrown whenever a SQLite error occurs.
     */
-    public func delete(db: Database) throws {
-        try withDataMapper { dataMapper in
-            try dataMapper.delete(db)
+    public func delete(db: Database) throws -> DeletionResult {
+        return try withDataMapper { dataMapper -> DeletionResult in
+            let changes = try dataMapper.delete(db)
             
             // Future calls to update will throw RowModelNotFound. Make the user
             // a favor and make sure this error is thrown even if she checks the
             // edited flag:
             edited = true
+            
+            if changes.changedRowCount > 0 {
+                return .RowDeleted
+            } else {
+                return .NoRowDeleted
+            }
         }
     }
     
@@ -328,7 +357,7 @@ public class RowModel {
               reloaded.
     */
     public func reload(db: Database) throws {
-        try withDataMapper { dataMapper in
+        try withDataMapper { dataMapper -> Void in
             let statement = try dataMapper.reloadStatement(db)
             if let row = statement.fetchOneRow() {
                 for (column, databaseValue) in row {
@@ -346,9 +375,9 @@ public class RowModel {
     // MARK: - DataMapper
     
     /// Creates a DataMapper, and translates DataMapperError into RowModelError.
-    private func withDataMapper(@noescape block: (DataMapper) throws -> Void) throws {
+    private func withDataMapper<R>(@noescape block: (DataMapper) throws -> R) throws -> R {
         do {
-            try block(DataMapper(self))
+            return try block(DataMapper(self))
         } catch let error as DataMapperError {
             switch error {
             case .RowNotFound:
@@ -457,32 +486,15 @@ public class RowModel {
         ///
         /// Returns (rowIDColumn, insertedRowID) if the row model has a
         /// currently nil RowID primary key, and nil otherwise.
-        func insert(db: Database) throws -> (String, Int64)? {
+        func insert(db: Database) throws -> UpdateStatement.Changes {
             // INSERT
             let insertStatement = try DataMapper.insertStatement(db, tableName: databaseTable.name, insertedColumns: Array(storedDatabaseDictionary.keys))
             let arguments = QueryArguments(storedDatabaseDictionary.values)
-            let changes = try insertStatement.execute(arguments: arguments)
-            
-            // Return inserted RowID column if needed: currently nil RowID primary key.
-            if let primaryKey = databaseTable.primaryKey, case .RowID(let rowIDColumn) = primaryKey {
-                guard let rowID = storedDatabaseDictionary[rowIDColumn] else {
-                    fatalError("\(rowModel.dynamicType).storedDatabaseDictionary must return the value for the primary key `(rowIDColumn)`")
-                }
-                if rowID == nil {
-                    // RowID is not set yet: tell RowModel
-                    return (rowIDColumn, changes.insertedRowID!)
-                } else {
-                    // RowID is already set: no need for RowID.
-                    return nil
-                }
-            } else {
-                // No RowID primary Key: no need for RowID
-                return nil
-            }
+            return try insertStatement.execute(arguments: arguments)
         }
         
         /// UPDATE
-        func update(db: Database) throws {
+        func update(db: Database) throws -> UpdateStatement.Changes {
             // Update requires strongPrimaryKeyDictionary
             guard let primaryKeyDictionary = strongPrimaryKeyDictionary else {
                 fatalError("Invalid primary key in \(rowModel)")
@@ -526,7 +538,7 @@ public class RowModel {
                 guard row != nil else {
                     throw DataMapperError.RowNotFound
                 }
-                return
+                return UpdateStatement.Changes(changedRowCount: 0, insertedRowID: nil)
             }
             
             // Update
@@ -538,10 +550,13 @@ public class RowModel {
             if changes.changedRowCount == 0 {
                 throw DataMapperError.RowNotFound
             }
+            
+            return changes
         }
         
         /// DELETE
-        func delete(db: Database) throws {
+        /// Returns whether a row was actually deleted
+        func delete(db: Database) throws -> UpdateStatement.Changes {
             // Delete requires strongPrimaryKeyDictionary
             guard let primaryKeyDictionary = strongPrimaryKeyDictionary else {
                 fatalError("Invalid primary key in \(rowModel)")
@@ -550,7 +565,7 @@ public class RowModel {
             // Delete
             let deleteStatement = try DataMapper.deleteStatement(db, tableName: databaseTable.name, conditionColumns: Array(primaryKeyDictionary.keys))
             let arguments = QueryArguments(primaryKeyDictionary.values)
-            try deleteStatement.execute(arguments: arguments)
+            return try deleteStatement.execute(arguments: arguments)
         }
         
         /// SELECT